Fix build warning in MM6X18542.p2 release
When building MM6X18542.p2 wifi release on ARM32 platforms, will meet
the following build warning, force the size_t type to be converted to
the unsigned long to avoid this warning.

mlinux/moal_proc.c: In function ‘woal_ssu_dump_read’:
/linux-lts-nxp/include/linux/kern_levels.h:5:25: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 2 has type ‘size_t’ {aka ‘unsigned int’} [-Wformat=]
    5 | #define KERN_SOH        "\001"          /* ASCII Start Of Header */
      |                         ^~~~~~

+++ mlinux/moal_proc.c
@@ -1645,7 +1645,7 @@ static int woal_ssu_dump_read(struct seq_file *sfp, void *data)
 	if (sfp->size < ((handle->ssu_dump_len * 9) / 4)) {
 		PRINTM(MCMND,
 		       "ssu dump size too big, size=%lu, ssu_dump_len=%lu\n",
-		       sfp->size,
+		       (unsigned long)sfp->size,
 		       (unsigned long)((handle->ssu_dump_len * 9) / 4));
 		sfp->count = sfp->size;
 		ret = 0;
